feat(src/util/dynamic_library): rename, and refactor to support multiple platforms
